Created the doc directory for user docs of the .pod and .kwid forms.
Started converting perlpod.pod to perlkwid.kwid. This is a great way to create
the kwid language and know whether it will work. My notes:

I am half way through this Pod to Kwid conversion. I think it is
going quite well. Kwid is much easier to read and write so far. But
there is a definite mess and smell to this POD stuff. Do I want to
support it all? Sleep will help me decide...

Some problems so far.

Kwiki lists are generally one line per item. Pod is not such. Nesting Pod
type lists makes it obvious why there is the whole:
  =over
  =item
  =item
  =back
mess. It is not optimized for simple lists, but handles non-bulleted,
not-numbered list like things. I will have to look at the generated HTML of
pod2html to grok what is really intended.

I imagine a solution like this:

.list
- widget()
The widget function does this...
- gadget()
The gadget function does that...
.list 2
- gadget(Int)
The int form does blah...
- gadget(Str)
The str form does blech...
.list 2.
.list.

Also the *bold* and /italic/ formats will have {*bold*} and {/italic/} forms
for when things get hairy...

More tommorrow...



git-svn-id: file:///data/svn/pugs@298 c213334d-75ef-0310-aa23-eaa082d1ae64