Test-Less comes to Pugs
Test::Less helps speed up development testing by allowing people to categorize
tests with various keywords, and then prove subsets of those tests.

It keeps all its info in t/Test-Less/index.txt. You don't have to use it if
you don't want to. It is an author-side tool, so support does not need to be
distributed. But the index should be distributed...




git-svn-id: file:///data/svn/pugs@4416 c213334d-75ef-0310-aa23-eaa082d1ae64