The PUGS_EMBED perl5 embedding avoidance flag is now "noperl5". Files: INSTALL, Makefile.PL
Perl5 was made default on 2006-08-15, and the "-perl5"/"!perl5" flags added.

However, Makefile.PL uses /\bFLAG\b/ tests.
The positive "perl5" flag continues to exist, and be used.
Tests for it match the new new, oppositely sensed, flags.
So much of Makefile was not honoring "-perl5" requests.

A new flag "noperl5" has now been added to replace "-perl5"/"!perl5".
Makefile does the right thing with it.
The flags "-perl5" and "!perl5" still exist, are depreciated,
generate a warning, and are still buggy.

"noperl5" has the advantage of matching the existing convention.
If we want a new convention, that's fine, but requires additional changes.

The depreciation code, and checks for [-!], should be removed for next release.


git-svn-id: file:///data/svn/pugs@13193 c213334d-75ef-0310-aa23-eaa082d1ae64