In t/ (mostly), converted eval_ok and eval_is to ok(eval and is(eval. The majority were incorrectly assuming lexical and package context were shared. Some knowingly depended on this, and will need to be changed "back" (but hopefully to either local helper code, or something with a less confusion creating name).
The quality of position information in error messages about the eval()ed string is reduced.  But without a good macro implementation available on all p6 implementations, this does not currently seem avoidable.

eval_dies_ok use was not touched, as it was being used correctly, and it would require a (very) slightly less trivial fix.

TreyHarris++ did a lot of related work.

eval_ok and eval_is are now gone from:
  t/ test files with the exception of t/02-test-pm/1-basic.t
  t/REAMDE
  ext/Set/t/overload.t


git-svn-id: file:///data/svn/pugs@13913 c213334d-75ef-0310-aa23-eaa082d1ae64