[kp6] Documented the meaning of p5landish
You have a Perl 6 string object that you want to compare using `eq`, forexample. The object itself would be a Perl 6 
object, and you would usep5landish on it to get at the actual string contained within it so you cancompare it.In 
short, p5landish returns the actual value contained in the object.

See: 
http://irclog.perlgeek.de/perl6/2007-11-27#i_152004



git-svn-id: file:///data/svn/pugs@18942 c213334d-75ef-0310-aa23-eaa082d1ae64