[kp6]
fglock suggested I make Makefile.PL more maintainable.  I believe that the first step should be to extract the 
configuration settings for each build out of the Makefile.PL to reduce FUD.  The code is fairly long and difficult to 
read; by extracting the code into a YAML (targets.yml) file, this simplifies the code and makes it easier to read.

The next step will be to sort out the pod creation.

dlocaus @ #perl6 irc.freenode.net



git-svn-id: file:///data/svn/pugs@19321 c213334d-75ef-0310-aa23-eaa082d1ae64